Support showing a specific referral
We add a /random route which redirects to the show route after picking
a random referral. We want referrals to be bookmarkable and shareable
so we need a consistent link to a specific referral.

To do this we need a unique id for a referral. We use a hash of the
service provider combined with the code for the id. We use a hash
because codes are often urls and create ugly paths when parameterised.